# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API para el agente de menú semanal
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RecipeCategory =
  (typeof RecipeCategory)[keyof typeof RecipeCategory];

export const RecipeCategory = {
  primero: "primero",
  segundo: "segundo",
  otro: "otro",
} as const;

export interface Recipe {
  id: number;
  name: string;
  category: RecipeCategory;
  ingredients: string[];
  instructions?: string;
  createdAt?: string;
}

export type CreateRecipeInputCategory =
  (typeof CreateRecipeInputCategory)[keyof typeof CreateRecipeInputCategory];

export const CreateRecipeInputCategory = {
  primero: "primero",
  segundo: "segundo",
  otro: "otro",
} as const;

export interface CreateRecipeInput {
  name: string;
  category: CreateRecipeInputCategory;
  ingredients: string[];
  instructions?: string;
}

export interface GenerateMenuInput {
  /** Preferencias o restricciones opcionales */
  preferences?: string;
  /** IDs de recetas a excluir */
  excludeRecipes?: number[];
  /** Número de días a planificar */
  daysCount?: number;
}

export interface MealPlan {
  primero?: Recipe;
  segundo?: Recipe;
}

export interface DayMenu {
  day: string;
  lunch: MealPlan;
  dinner: MealPlan;
}

export interface WeeklyMenu {
  id: number;
  days: DayMenu[];
  createdAt?: string;
}

export interface GenerateShoppingListInput {
  menuId: number;
}

export interface MercadonaProduct {
  id: string;
  name: string;
  price?: number;
  unit?: string;
  thumbnail?: string;
  ingredient: string;
}

export interface ShoppingListItem {
  ingredient: string;
  recipes: string[];
  mercadonaProducts?: MercadonaProduct[];
}

export interface ShoppingList {
  menuId: number;
  items: ShoppingListItem[];
  generatedAt?: string;
}

export interface MercadonaSearchInput {
  ingredients: string[];
}

export interface MercadonaSearchResult {
  results: ShoppingListItem[];
}

export type ListRecipesParams = {
  category?: string;
};