# lib/api-zod/src/generated/types/createRecipeInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API para el agente de menú semanal
 * OpenAPI spec version: 0.1.0
 */
import type { CreateRecipeInputCategory } from "./createRecipeInputCategory";

export interface CreateRecipeInput {
  name: string;
  category: CreateRecipeInputCategory;
  ingredients: string[];
  instructions?: string;
}