# lib/api-zod/src/generated/types/dayMenu.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API para el agente de menú semanal
 * OpenAPI spec version: 0.1.0
 */
import type { MealPlan } from "./mealPlan";

export interface DayMenu {
  day: string;
  lunch: MealPlan;
  dinner: MealPlan;
}