# lib/api-zod/src/generated/types/generateMenuInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API para el agente de menú semanal
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateMenuInput {
  /** Preferencias o restricciones opcionales */
  preferences?: string;
  /** IDs de recetas a excluir */
  excludeRecipes?: number[];
  /** Número de días a planificar */
  daysCount?: number;
}