# lib/api-zod/src/generated/types/mealPlan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API para el agente de menú semanal
 * OpenAPI spec version: 0.1.0
 */
import type { Recipe } from "./recipe";

export interface MealPlan {
  primero?: Recipe;
  segundo?: Recipe;
}