# lib/api-zod/src/generated/types/mercadonaProduct.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API para el agente de menú semanal
 * OpenAPI spec version: 0.1.0
 */

export interface MercadonaProduct {
  id: string;
  name: string;
  price?: number;
  unit?: string;
  thumbnail?: string;
  ingredient: string;
}