# lib/api-zod/src/generated/types/mercadonaSearchResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API para el agente de menú semanal
 * OpenAPI spec version: 0.1.0
 */
import type { ShoppingListItem } from "./shoppingListItem";

export interface MercadonaSearchResult {
  results: ShoppingListItem[];
}